Added Saas, reordered components
