+++ package.json
@@ -35,7 +35,7 @@
   },
   "devDependencies": {
     "lato-font": "^3.0.0",
-    "node-sass": "^4.13.1",
+    "node-sass": "^4.14.0",
     "sass-loader": "^8.0.2"
   }
 }
