+++ src/App.js
@@ -1,8 +1,7 @@
-import React, { Component } from 'react';
-import './App.css';
+import React, { Component } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
-import Home from './pages/Home'
+import Home from "./components/pages/Home";
 
 class App extends Component {
   render() {
