+++ src/assets/images/Polygon_5_jaune.png
