+++ src/assets/images/Polygon_5_vert.png
