+++ src/assets/images/kaino_logo_blanc_text.png
