+++ src/assets/style/components/_header.scss
@@ -1,0 +1,6 @@
+.logo {
+  animation: App-logo-spin infinite 20s linear;
+  height: 80px;
+  margin: 1vw;
+  transition: all 0.3s ease 0s;
+}
