+++ src/assets/style/components/_home.scss
@@ -1,0 +1,26 @@
+:root {
+  --bgcolor: #f7f2ff;
+}
+
+html,
+body {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  background-color: var(--bgcolor);
+  background-attachment: fixed; /* Le fond restera fixe */
+  font-size: 100%;
+}
+
+h1 {
+  color: white;
+  margin-top: 0;
+  font-family: Lato;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 24px;
+  line-height: 54px;
+  letter-spacing: 0.02em;
+}
