+++ src/assets/style/organisms/_concept.scss
@@ -1,0 +1,36 @@
+.concept {
+  position: relative;
+  width: 100%;
+  height: auto;
+  left: 0px;
+  right: 0px;
+}
+
+.hex3 {
+  position: absolute;
+  width: 50%;
+  height: auto;
+  top: -80px;
+  left: 0;
+}
+
+.titreSection {
+  position: relative;
+  width: 90%;
+  font-family: "MaculaLine";
+  font-size: 40px;
+  text-align: right;
+  color: #5832a0;
+  margin-top: 100px;
+}
+
+.concept3en1 {
+  position: relative;
+  height: auto;
+  margin-top: 20px;
+  width: 100%;
+}
+.pictoSize {
+  width: auto;
+  height: 50%;
+}
