+++ src/assets/style/organisms/_contact.scss
@@ -1,0 +1,24 @@
+.contact {
+  position: relative;
+  width: 100%;
+  height: auto;
+  left: 0px;
+  right: 0px;
+  top: 100px;
+}
+
+.hexVert {
+  position: absolute;
+  width: 50%;
+  height: auto;
+  left: 0;
+}
+
+.boxcontact {
+  position: relative;
+  width: 344px;
+  height: 512px;
+  left: 2%;
+  top: 20%;
+  border: solid;
+}
