+++ src/assets/style/organisms/_footer.scss
@@ -1,0 +1,15 @@
+.footer {
+  position: relative;
+  width: 100%;
+  height: 150px;
+  bottom: 0px;
+  background-color: #b2d8f6;
+  margin-top: 200px;
+}
+
+.txtcopyright {
+  position: relative;
+  width: 100%;
+  color: white;
+  vertical-align: bottom;
+}
