+++ src/assets/style/organisms/_header.scss
@@ -1,0 +1,97 @@
+.entete {
+  position: relative;
+  width: 100%;
+  height: auto;
+  left: 0px;
+  right: 0px;
+  top: 0px;
+  background: #f7f2ff;
+  margin-bottom: 10%;
+}
+.polygone1 {
+  width: 100%;
+  height: 100%;
+}
+
+.bigTitre {
+  position: absolute;
+  width: 35%;
+  left: 0;
+  right: 0;
+  margin-left: auto;
+  margin-right: auto;
+  top: 8%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.img_titre {
+  width: 100%;
+  height: auto;
+  margin-top: 40pt;
+  margin-bottom: 5pt;
+}
+.boxSousTitre {
+  width: 100%;
+  font-size: 100% !important;
+}
+
+.boxderoulant {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.flechederoulant {
+  position: relative;
+  padding: 0.5em;
+}
+
+.petitBadge1 {
+  position: absolute;
+  width: 100%;
+  top: 48%;
+  height: auto;
+}
+.hex {
+  position: absolute;
+  width: 25%;
+  height: auto;
+}
+.hex-1 {
+  left: -30px;
+}
+.hex-2 {
+  left: 60%;
+  margin-top: 8%;
+}
+
+.fleche {
+  position: absolute;
+  width: auto;
+  height: auto;
+}
+
+.position123 {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 25%;
+  height: auto;
+  left: -30px;
+}
+
+.hexa123 {
+  position: relative;
+  width: 100%;
+  height: auto;
+}
+.hexa456 {
+  position: absolute;
+  width: auto;
+  height: auto;
+}
