+++ src/assets/style/organisms/_menu.scss
@@ -1,0 +1,24 @@
+.menu {
+  position: absolute;
+  width: 100%;
+  height: auto;
+  top: 16px;
+}
+
+.navigationBar {
+  font-size: 1em;
+  position: absolute;
+  top: 0;
+  right: 2%;
+  color: white;
+
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  margin: 0.5em;
+}
+
+.navigationItem {
+  padding-left: 0.5em;
+  padding-right: 0.5em;
+}
