+++ src/assets/style/organisms/_offers.scss
@@ -1,0 +1,15 @@
+.offres {
+  position: relative;
+  width: 100%;
+  height: auto;
+  left: 0px;
+  right: 0px;
+  top: 100px;
+}
+
+.kkk_size {
+  position: absolute;
+  width: 100%;
+  top: 250px;
+  left: 0;
+}
