+++ src/assets/style/organisms/_partners.scss
@@ -1,0 +1,19 @@
+.partenaires {
+  position: relative;
+  width: 100%;
+  height: auto;
+  left: 0px;
+  right: 0px;
+  top: 100px;
+}
+
+.cardlogo {
+  position: relative;
+  display: inline-block;
+  margin: 1rem;
+  width: 300px;
+  height: auto;
+}
+.ttt {
+  width: 200px;
+}
