+++ src/assets/style/utilities/_colors.sass
@@ -1,0 +1,2 @@
+
+$background: #1f29ac
