+++ src/assets/style/utilities/_fonts.sass
@@ -1,0 +1,11 @@
+@font-face
+    font-family: "MaculaSolid", Lato
+    src: url("../../fonts/MaculaSolid.ttf") format("truetype")
+
+    &
+        font-family: "MaculaLine", Lato
+        src: url("../../fonts/MaculaLine.ttf") format("truetype")
+
+    &
+        font-family: "MaculaShaded", Lato
+        src: url("../../fonts/MaculaShaded.ttf") format("truetype")
