--- src/component/Header.css
@@ -1,6 +1,0 @@
-.logo {
-    animation: App-logo-spin infinite 20s linear;
-    height: 80px;
-    margin: 1vw;
-    transition: all 0.3s ease 0s;
-}
