--- src/component/Header.js
@@ -1,16 +1,0 @@
-import React, { Fragment, useState } from "react";
-import './Header.css'
-import logo from '../img/logo-site.png';
-import { Link } from 'react-router-dom';
-
-const Header = ({}) => {
-
-    return (
-        <div>
-            <p>Hello World</p>
-        </div>
-
-    )
-}
-
-export default Header
