--- src/component/test.css
@@ -1,42 +1,0 @@
-.header{
-    position: absolute;
-    width: 100%;
-    height: auto;
-    left: 0px;
-    top: 0px;
-
-    background: #F7F2FF;
-    }
-
-.menu{
-    position: absolute;
-    width: 100%;
-    height: auto;
-    left: 11px;
-    top: 16px;
-    }
-
-.h3test {
-    position: absolute;
-    left: 47.12%;
-    right: 40.42%;
-    top: 19.39%;
-    bottom: 43.88%;
-    }
-
-.polygone1 {
-    width: 100%;
-    height: 100%;
-    }
-
-.bigTitre {
-    position: absolute;
-    width:100%;
-    top: 8%;
-    }
-.img_titre {
-    width: 45%;
-    height: auto;
-    margin-top: 40pt;
-    margin-bottom: 30pt;
-    }
