+++ src/components/organisms/Footer.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+class Footer extends React.Component {
+  render() {
+    return (
+      <div className="footer">
+        <div>
+          <p className="txtcopyright">
+            Copyright © 2020 Kaïno – Security Bootcamp – escape room
+          </p>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Footer;
